So the issue of the weird war structure is resolved by removing the archive classes thingy from the maven war plugin, so that's that.
After this change, the server logs state that weld is properly managing the CDI beans that are meant to be injected, after having them implement Serializable that is.

The issue of the injection null pointer exception still persists, and the exception seems to be too short to be useful, I am still working on reducing the application to a
reasonable setup. Mainly, I need to deal with the WEB-INF/lib directory containing jetty specific jars, that should only be present outside the webapp's WEB-INF.
After resolving this issue, I will pay a closer attention to the null pointer's legendary nuisance.